requirements.txt
streamlit
langchain-mcp-adapters
langchain-openai
langchain-core
openai
python-dotenv
nest_asyncio
